requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-environ>=0.11.0
pymysql>=1.1.0
django-cors-headers>=4.3.0
pandas>=2.0.0
openpyxl>=3.1.0
